Split listen model classes to separate files
